+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- docker build stage: revert user after applying `chown` workaround for inter-stage copy
 
 ## [0.4.0] - 2020-09-10
 ### Added